requirements.txt
Django==3.2.5
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.1
